compiler: save integer as a string instead of long
git-svn-id: http://asn1-compiler.googlecode.com/svn/trunk@111 b35bc7a6-0bd6-6437-dd7b-65e59c63c965

+++ compiler/src/parser/AsnParser.java
@@ -1564,7 +1564,7 @@ try {Token tok;
     jjtree.closeNodeScope(jjtn000, true);
     jjtc000 = false;
     jjtreeCloseNodeScope(jjtn000);
-     jjtn000.setNumber(Long.valueOf(((null != m) ? m.toString() : "") + n.toString()).longValue());
+     jjtn000.setNumber(((null != m) ? m.toString() : "") + n.toString());
     } finally {
     if (jjtc000) {
       jjtree.closeNodeScope(jjtn000, true);
